requirements.txt
streamlit==1.33.0
tensorflow>=2.15.0
opencv-python-headless==4.9.0.80
numpy
Pillow
